Closes #1890 and #1900 to properly check for dates and fix other bugs in checkSpec

+++ inst/examples/2_AdverseEventWorkflow.R
@@ -29,7 +29,7 @@ spec:
       type: character
     timeonstudy:
       required: true
-      type: numeric
+      type: integer
 steps:
   - output: vThreshold
     name: ParseThreshold
